requirements.txt
#python==3.10.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
pandas
